Move benchmaker to a separate submodule
