+++ start/common/src/main/resources/config/spec-constants.yml
