+++ start/simulator/src/main/java/org/ethereum/beacon/simulator/SimulatorLauncher.java
@@ -51,7 +51,7 @@ import org.ethereum.beacon.schedulers.LoggerMDCExecutor;
 import org.ethereum.beacon.schedulers.Schedulers;
 import org.ethereum.beacon.schedulers.TimeController;
 import org.ethereum.beacon.schedulers.TimeControllerImpl;
-import org.ethereum.beacon.simulator.util.SimulateUtils;
+import org.ethereum.beacon.util.SimulateUtils;
 import org.ethereum.beacon.validator.crypto.BLS381Credentials;
 import org.ethereum.beacon.wire.LocalWireHub;
 import org.ethereum.beacon.wire.WireApi;
@@ -452,7 +452,8 @@ public class SimulatorLauncher implements Runnable {
       SimulationPlan simulationPlan = (SimulationPlan) config.getPlan();
 
       ConfigBuilder<SpecData> specConfigBuilder =
-          new ConfigBuilder<>(SpecData.class).addYamlConfigFromResources("/config/spec-constants.yml");
+          new ConfigBuilder<>(SpecData.class).addYamlConfigFromResources(
+              "/config/spec-constants.yml");
       if (config.getChainSpec().isDefined()) {
         specConfigBuilder.addConfig(config.getChainSpec());
       }
